requirements.txt
aiodns==3.2.0
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.2
attrs==25.1.0
bidict==0.23.1
blinker==1.9.0
ccxt==4.4.52
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.0
Flask==3.1.0
Flask-Login==0.6.3
Flask-SocketIO==5.5.1
frozenlist==1.5.0
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
MarkupSafe==3.0.2
multidict==6.1.0
narwhals==1.24.1
numpy==2.2.2
packaging==24.2
pandas==2.2.3
pandas_ta==0.3.14b0
plotly==6.0.0
propcache==0.2.1
pycares==4.5.0
pycparser==2.22
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.11.2
python-socketio==5.12.1
pytz==2024.2
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.1
setuptools==75.8.0
simple-websocket==1.1.0
six==1.17.0
tailwindcss==0.0.1
threadpoolctl==3.5.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
Werkzeug==3.1.3
wsproto==1.2.0
xgboost==2.1.3
yarl==1.18.3
